move user and store directory into auth package
