+++ auth/google/google.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 
 	"github.com/coreos/go-oidc"
-	"github.com/pankona/hashira/user"
+	"github.com/pankona/hashira/auth/user"
 	uuid "github.com/satori/go.uuid"
 	"golang.org/x/net/context"
 	"golang.org/x/oauth2"
