+++ auth/store/dsstore.go
